Add WealthGroupCharacteristicValue.source field - see HEA-144
The BSS 'WB' worksheet, which contains three types of  Wealth Group
Characteristic Value.

The first set of columns contain values from the Community Interviews
(Form 3). The second set of columns contain values from the Wealth Group
Interviews in each Community (Form 4). The third set of columns contain
the summary value and range that is considered representative for the
Baseline, and is derived from the Community and Wealth Group Interviews
as part of the baseline process.

Prior to this commit, the WealthGroupCharacteristicValue was either for
a BaselineWealthGroup, in which case it was from the Summary/From/To set
of columns. Or it was for a CommunityWealthGroup, in which case we were
assuming that it was the value that the Wealth Group in the Community
had reported - i.e. the value from the Form 4 Wealth Group Interview.
However, this doesn't leave us any way to store the Form 3 Community
Interview values, which are in the first set of columns, e.g. C:M, which
come from the discussion with the community leaders, but are still a
value for each wealth group.

Therefore, we add a `source` field to `WealthGroupCharacteristicValue`
which allows us to capture both the Form 3 value and the Form 4 value
and distinguish them from one another. We also use `source` to indicate
a Summary value, which makes filtering easier than using
`.filter(wealth_group__community__isnull=True)`.