Replace model_utils autofields with Django equivalents - see HEA-98
HEA uses a TimeStampedModel to automatically include
`created` and `modified` fields in each model. Previously,
it did this using django_model_utils TimeStampedModel,
but there has been equivalent functionality in Django for
many years, using `DateTimeField(auto_add=True)`. The
Django functionality has the added advantage that the
field can be detected as automaintained, and therefore
ignored by the ingestion pipelines.

This commit changes the fields to use the Django ones.
It also changes all the migrations where in introduced a
new model that included those fields. Although the
migration has changed, it doesn't need to be re-run,
because the underlying column in the database has not
changed.