Make MilkProduction attributes optional - see HEA-98
Some BSS record that the Wealth Group had 0 milking animals.
In this case the `milking_animals` contains 0 but the `lactation_days`
and `daily_production` are blank. Therefore, these fields need to
be `null=True`.

I have added validation to `MilkProduction.clean()` to check that
the attributes have values if `milking_animals` has a value.