Extend get_summary_bss_label_dataframe to include metadata - see HEA-214
Extend get_summary_bss_label_dataframe so that it reads
the ActivityLabel and WealthCharacteristicLabel instances
and joins the appropriate metadata attributes to the label
dataframe, such that the result can be used to recreate the
corresponding worksheet from the Reference Data GSheet.

This includes using the Google Translate API to translate
non-English labels.