Support product__name attribute in Livelihood Activity instances - see HEA-214
Some BSS incorrectly specify the product in the value columns instead of in the label column. Therefore, if we have specified the product__name as the attribute, check that the product can be identified and is the same for all columns and then add it to the Livelihood Strategy. This avoids needing to create manual corrections
for all the BSSs where the product has been specified in the value
columns.