Better calculation of all_label_attributes - see HEA-327
Moves the calculation of `all_label_attributes` into a
separate function, so that it can be used to test the behavior of
a label in a Jupyter Notebook, for example.   It also changes the
SeasonLookup and UnitOfMeasureLookup so that they raise an
error if they find multiple matches, but still ignore lookups that
don't find any matches.