Fix bug in get_wealth_group_dataframe for WB - see HEA-573
Francophone BSS have `à` in row 4 in the WB worksheet in the summary columns
(`synthèse | de | à`). And `a` is an alias for the B/O Wealth Group Category (from
Aisé). So the previous code was matching that and thinking the data was for B/O
instead of part of the WB summary.