Update livelihood_activity_label_recognition_dataframe - see HEA-807
`get_all_label_attributes` now conditionally recognizes labels
that only contain a product name or alias. This means that we
have labels where the attributes don't come from either the
regexes or the override activity labels. Therefore, we need to
update `livelihood_activity_label_recognition_dataframe` so
that it shows the set of recogmized attributes as well as the
regex and override attributes.