Better factories - see HEA-659
The previous set of LivelihoodActivityFactory subclasses created
all attributes for all Strategy Types. This is inaccurate - the production
Strategies don't ever contain quantity_purchased or expenditure, and
the OtherCashIncome and OtherPurchase Strategies contain either
income or expenditure respectively, and not quantity_produced,
kcals_consumed, etc.  To make the test data more realistic we have
updated the subclasses so that they produce realistic data for the
Strategy.  This also required updating some tests that assumed that
all attributes would be populated for all subclasses.