Support ordering by related fields in AggregatingViewSet - see HEA-659
Before this commit the fields where mapped to the database by a custom function
(`field_to_database_path`) but the fields returned by `serializer.Meta.get_fields()`
wasn't aware of the actual paths. Consequently the OrderingFilter wouldn't recognize
them and we couldn't sort by them.  The solution is that we will require all related
fields to be defined using a new `AggregatingViewSet.get_queryset_annotations()`
so that all fields are available directly on the queryset. This actually makes the
implmentation simpler.